refactor(api): :art: Move generateMovieTitles into movieTitle factory
